no google classification in analyzer
