+++ pkg/fanal/licensing/classifier_test.go
